+++ includes/frontend/class-loop.php
@@ -127,7 +127,7 @@ class Reykjavik_Loop {
 					$current = ( get_query_var( 'paged' ) ) ? ( absint( get_query_var( 'paged' ) ) ) : ( 1 );
 
 					$output = '<nav class="pagination" role="navigation" aria-labelledby="pagination-label" data-current="' . esc_attr( $current ) . '" data-total="' . esc_attr( $total ) . '">'
-					          . '<h2 class="screen-reader-text" id="pagination-label">' . esc_attr__( 'Posts Navigation', 'reykjavik' ) . '</h2>'
+					          . '<h2 class="screen-reader-text" id="pagination-label">' . esc_html__( 'Posts Navigation', 'reykjavik' ) . '</h2>'
 					          . $output
 					          . '</nav>';
 				}